editoast: move schema/operation to editoast::infra_cache
