+++ editoast/src/generated_data/electrification.rs
@@ -12,6 +12,7 @@ use super::utils::InvolvedObjects;
 use super::GeneratedData;
 use crate::diesel::ExpressionMethods;
 use crate::error::Result;
+use crate::infra_cache::operation::CacheOperation;
 use crate::infra_cache::InfraCache;
 use crate::schema::ObjectType;
 use crate::tables::infra_layer_electrification::dsl;
@@ -39,8 +40,8 @@ impl GeneratedData for ElectrificationLayer {
     async fn update(
         conn: &mut PgConnection,
         infra: i64,
-        operations: &[crate::schema::operation::CacheOperation],
-        infra_cache: &crate::infra_cache::InfraCache,
+        operations: &[CacheOperation],
+        infra_cache: &InfraCache,
     ) -> Result<()> {
         let involved_objects =
             InvolvedObjects::from_operations(operations, infra_cache, ObjectType::Electrification);
