+++ editoast/src/generated_data/neutral_section.rs
@@ -6,6 +6,7 @@ use diesel_async::RunQueryDsl;
 
 use super::GeneratedData;
 use crate::error::Result;
+use crate::infra_cache::operation::CacheOperation;
 use crate::infra_cache::InfraCache;
 
 pub struct NeutralSectionLayer;
@@ -27,8 +28,8 @@ impl GeneratedData for NeutralSectionLayer {
     async fn update(
         _conn: &mut PgConnection,
         _infra: i64,
-        _operations: &[crate::schema::operation::CacheOperation],
-        _infra_cache: &crate::infra_cache::InfraCache,
+        _operations: &[CacheOperation],
+        _infra_cache: &InfraCache,
     ) -> Result<()> {
         // TODO: we don't manage the update of the neutral_section layer for the moment
         Ok(())
