+++ editoast/src/views/infra/auto_fixes/mod.rs
@@ -15,15 +15,15 @@ use tracing::error;
 use crate::error::InternalError;
 use crate::error::Result;
 use crate::generated_data::generate_infra_errors;
+use crate::infra_cache::operation::CacheOperation;
+use crate::infra_cache::operation::DeleteOperation;
+use crate::infra_cache::operation::Operation;
+use crate::infra_cache::operation::RailjsonObject;
+use crate::infra_cache::operation::UpdateOperation;
 use crate::infra_cache::InfraCache;
 use crate::infra_cache::ObjectCache;
 use crate::modelsv2::prelude::*;
 use crate::modelsv2::Infra;
-use crate::schema::operation::CacheOperation;
-use crate::schema::operation::DeleteOperation;
-use crate::schema::operation::Operation;
-use crate::schema::operation::RailjsonObject;
-use crate::schema::operation::UpdateOperation;
 use crate::schema::InfraError;
 use crate::schema::OSRDIdentified as _;
 use crate::schema::OSRDObject;
@@ -329,10 +329,10 @@ mod tests {
     use crate::fixtures::tests::db_pool;
     use crate::fixtures::tests::empty_infra;
     use crate::fixtures::tests::small_infra;
+    use crate::infra_cache::operation::DeleteOperation;
+    use crate::infra_cache::operation::Operation;
+    use crate::infra_cache::operation::RailjsonObject;
     use crate::infra_cache::InfraCacheEditoastError;
-    use crate::schema::operation::DeleteOperation;
-    use crate::schema::operation::Operation;
-    use crate::schema::operation::RailjsonObject;
     use crate::schema::utils::Identifier;
     use crate::schema::ApplicableDirectionsTrackRange;
     use crate::schema::BufferStop;
